addr id size

+++ src/addrman.cpp
@@ -109,7 +109,7 @@ void CAddrMan::RemoveInvalid()
     }
 }
 
-CAddrInfo* CAddrMan::Find(const CNetAddr& addr, int* pnId)
+CAddrInfo* CAddrMan::Find(const CNetAddr& addr, int64_t* pnId)
 {
     AssertLockHeld(cs);
 
@@ -124,11 +124,11 @@ CAddrInfo* CAddrMan::Find(const CNetAddr& addr, int* pnId)
     return nullptr;
 }
 
-CAddrInfo* CAddrMan::Create(const CAddress& addr, const CNetAddr& addrSource, int* pnId)
+CAddrInfo* CAddrMan::Create(const CAddress& addr, const CNetAddr& addrSource, int64_t* pnId)
 {
     AssertLockHeld(cs);
 
-    int nId = nIdCount++;
+    int64_t nId = nIdCount++;
     mapInfo[nId] = CAddrInfo(addr, addrSource);
     mapAddr[addr] = nId;
     mapInfo[nId].nRandomPos = vRandom.size();
@@ -147,8 +147,8 @@ void CAddrMan::SwapRandom(unsigned int nRndPos1, unsigned int nRndPos2)
 
     assert(nRndPos1 < vRandom.size() && nRndPos2 < vRandom.size());
 
-    int nId1 = vRandom[nRndPos1];
-    int nId2 = vRandom[nRndPos2];
+    int64_t nId1 = vRandom[nRndPos1];
+    int64_t nId2 = vRandom[nRndPos2];
 
     assert(mapInfo.count(nId1) == 1);
     assert(mapInfo.count(nId2) == 1);
@@ -160,7 +160,7 @@ void CAddrMan::SwapRandom(unsigned int nRndPos1, unsigned int nRndPos2)
     vRandom[nRndPos2] = nId1;
 }
 
-void CAddrMan::Delete(int nId)
+void CAddrMan::Delete(int64_t nId)
 {
     AssertLockHeld(cs);
 
@@ -182,7 +182,7 @@ void CAddrMan::ClearNew(int nUBucket, int nUBucketPos)
 
     // if there is an entry in the specified bucket, delete it.
     if (vvNew[nUBucket][nUBucketPos] != -1) {
-        int nIdDelete = vvNew[nUBucket][nUBucketPos];
+        int64_t nIdDelete = vvNew[nUBucket][nUBucketPos];
         CAddrInfo& infoDelete = mapInfo[nIdDelete];
         assert(infoDelete.nRefCount > 0);
         infoDelete.nRefCount--;
@@ -193,7 +193,7 @@ void CAddrMan::ClearNew(int nUBucket, int nUBucketPos)
     }
 }
 
-void CAddrMan::MakeTried(CAddrInfo& info, int nId)
+void CAddrMan::MakeTried(CAddrInfo& info, int64_t nId)
 {
     AssertLockHeld(cs);
 
@@ -216,7 +216,7 @@ void CAddrMan::MakeTried(CAddrInfo& info, int nId)
     // first make space to add it (the existing tried entry there is moved to new, deleting whatever is there).
     if (vvTried[nKBucket][nKBucketPos] != -1) {
         // find an item to evict
-        int nIdEvict = vvTried[nKBucket][nKBucketPos];
+        int64_t nIdEvict = vvTried[nKBucket][nKBucketPos];
         assert(mapInfo.count(nIdEvict) == 1);
         CAddrInfo& infoOld = mapInfo[nIdEvict];
 
@@ -247,7 +247,7 @@ void CAddrMan::Good_(const CService& addr, bool test_before_evict, int64_t nTime
 {
     AssertLockHeld(cs);
 
-    int nId;
+    int64_t nId;
 
     nLastGood = nTime;
 
@@ -319,7 +319,7 @@ bool CAddrMan::Add_(const CAddress& addr, const CNetAddr& source, int64_t nTimeP
         return false;
 
     bool fNew = false;
-    int nId;
+    int64_t nId;
     CAddrInfo* pinfo = Find(addr, &nId);
 
     // Do not set a penalty for a source's self-announcement
@@ -432,7 +432,7 @@ CAddrInfo CAddrMan::Select_(bool newOnly)
                 nKBucket = (nKBucket + insecure_rand.randbits(ADDRMAN_TRIED_BUCKET_COUNT_LOG2)) % ADDRMAN_TRIED_BUCKET_COUNT;
                 nKBucketPos = (nKBucketPos + insecure_rand.randbits(ADDRMAN_BUCKET_SIZE_LOG2)) % ADDRMAN_BUCKET_SIZE;
             }
-            int nId = vvTried[nKBucket][nKBucketPos];
+            int64_t nId = vvTried[nKBucket][nKBucketPos];
             assert(mapInfo.count(nId) == 1);
             CAddrInfo& info = mapInfo[nId];
             if (insecure_rand.randbits(30) < fChanceFactor * info.GetChance() * (1 << 30))
@@ -449,7 +449,7 @@ CAddrInfo CAddrMan::Select_(bool newOnly)
                 nUBucket = (nUBucket + insecure_rand.randbits(ADDRMAN_NEW_BUCKET_COUNT_LOG2)) % ADDRMAN_NEW_BUCKET_COUNT;
                 nUBucketPos = (nUBucketPos + insecure_rand.randbits(ADDRMAN_BUCKET_SIZE_LOG2)) % ADDRMAN_BUCKET_SIZE;
             }
-            int nId = vvNew[nUBucket][nUBucketPos];
+            int64_t nId = vvNew[nUBucket][nUBucketPos];
             assert(mapInfo.count(nId) == 1);
             CAddrInfo& info = mapInfo[nId];
             if (insecure_rand.randbits(30) < fChanceFactor * info.GetChance() * (1 << 30))
@@ -464,14 +464,14 @@ int CAddrMan::Check_()
 {
     AssertLockHeld(cs);
 
-    std::unordered_set<int> setTried;
-    std::unordered_map<int, int> mapNew;
+    std::unordered_set<int64_t> setTried;
+    std::unordered_map<int64_t, int> mapNew;
 
     if (vRandom.size() != (size_t)(nTried + nNew))
         return -7;
 
     for (const auto& entry : mapInfo) {
-        int n = entry.first;
+        int64_t n = entry.first;
         const CAddrInfo& info = entry.second;
         if (info.fInTried) {
             if (!info.nLastSuccess)
@@ -619,8 +619,8 @@ void CAddrMan::ResolveCollisions_()
 {
     AssertLockHeld(cs);
 
-    for (std::set<int>::iterator it = m_tried_collisions.begin(); it != m_tried_collisions.end();) {
-        int id_new = *it;
+    for (std::set<int64_t>::iterator it = m_tried_collisions.begin(); it != m_tried_collisions.end();) {
+        int64_t id_new = *it;
 
         bool erase_collision = false;
 
@@ -638,7 +638,7 @@ void CAddrMan::ResolveCollisions_()
             } else if (vvTried[tried_bucket][tried_bucket_pos] != -1) { // The position in the tried bucket is not empty
 
                 // Get the to-be-evicted address that is being tested
-                int id_old = vvTried[tried_bucket][tried_bucket_pos];
+                int64_t id_old = vvTried[tried_bucket][tried_bucket_pos];
                 CAddrInfo& info_old = mapInfo[id_old];
 
                 // Has successfully connected in last X hours
@@ -682,11 +682,11 @@ CAddrInfo CAddrMan::SelectTriedCollision_()
 
     if (m_tried_collisions.size() == 0) return CAddrInfo();
 
-    std::set<int>::iterator it = m_tried_collisions.begin();
+    std::set<int64_t>::iterator it = m_tried_collisions.begin();
 
     // Selects a random element from m_tried_collisions
     std::advance(it, insecure_rand.randrange(m_tried_collisions.size()));
-    int id_new = *it;
+    int64_t id_new = *it;
 
     // If id_new not found in mapInfo remove it from m_tried_collisions
     if (mapInfo.count(id_new) != 1) {
@@ -700,7 +700,7 @@ CAddrInfo CAddrMan::SelectTriedCollision_()
     int tried_bucket = newInfo.GetTriedBucket(nKey, m_asmap);
     int tried_bucket_pos = newInfo.GetBucketPosition(nKey, false, tried_bucket);
 
-    int id_old = vvTried[tried_bucket][tried_bucket_pos];
+    int64_t id_old = vvTried[tried_bucket][tried_bucket_pos];
 
     return mapInfo[id_old];
 }
